moved download functionality to redux
